requirements.txt
cloudinary
pandas
torch
scipy
transformers
requests
chardet
neo4j
bs4
tabulate
textblob
Flask
matplotlib
git+https://github.com/JustAnotherArchivist/snscrape.git
gunicorn
python-dotenv